Résolution problème vuihandler
Former-commit-id: 41f342b825168e6b736d585a8a0d5a6e414cf5a4

+++ ME/base/so3/soo/include/soo/dev/vuihandler.h
@@ -73,8 +73,8 @@ typedef struct __attribute__((packed)) {
 
 typedef struct {
 	uint32_t		id;
-	uint8_t 		type;
-	size_t			size;
+	uint32_t 		type;
+	uint32_t			size;
 	uint8_t 		buf[VUIHANDLER_MAX_PAYLOAD_SIZE];
 } vuihandler_tx_request_t;
 
@@ -92,8 +92,8 @@ typedef struct {
 
 typedef struct {
 	uint32_t		id;
-	size_t			size;
-	uint8_t			type;
+	uint32_t			size;
+	uint32_t			type;
 	uint8_t 		buf[VUIHANDLER_MAX_PAYLOAD_SIZE];
 } vuihandler_rx_response_t;
